+++ src/components/ControlBar/ControlBar.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types'
 import cx from 'classnames'
 import classes from './styles/ControlBar.module.css'
 
+//Matches the height of .dragHandle in ControlBar.module.css
 export const DRAG_HANDLE_HEIGHT = 7
 
 class ControlBar extends React.Component {
@@ -50,10 +51,9 @@ class ControlBar extends React.Component {
     renderDragHandle = () =>
         typeof this.props.onChangeHeight === 'function' && (
             <div
-                data-testid="controlbar-drag-handle"
                 className={classes.draghandle}
-                style={{ height: DRAG_HANDLE_HEIGHT }}
                 onMouseDown={this.onStartDrag}
+                data-testid="controlbar-drag-handle"
             />
         )