+++ src/components/ControlBar/Filter.js
@@ -28,6 +28,7 @@ export const FilterUnconnected = ({
     onToggleMaxHeight,
 }) => {
     const [focusedClassName, setFocusedClassName] = useState('')
+    const [inputFocused, setInputFocus] = useState(false)
     const { width } = useWindowDimensions()
 
     const setFilterValue = event => {
@@ -59,8 +60,19 @@ export const FilterUnconnected = ({
         setFocusedClassName('')
     }
 
+    const onFocusInput = input => {
+        if (input && inputFocused && isSmallScreen(width)) {
+            return input.focus()
+        }
+    }
+
+    const toggleMaxHeight = () => {
+        onToggleMaxHeight()
+        setInputFocus(true)
+    }
+
     return isSmallScreen(width) && !isMaxHeight ? (
-        <button className={classes.searchButton} onClick={onToggleMaxHeight}>
+        <button className={classes.searchButton} onClick={toggleMaxHeight}>
             <SearchIcon className={classes.searchIcon} />
         </button>
     ) : (
@@ -69,7 +81,10 @@ export const FilterUnconnected = ({
             onFocus={onFocus}
             onBlur={onBlur}
         >
-            <SearchIcon className={classes.searchIcon} />
+            <SearchIcon
+                className={classes.searchIcon}
+                small={isSmallScreen(width) && isMaxHeight}
+            />
             <input
                 className={classes.input}
                 type="text"
@@ -78,6 +93,7 @@ export const FilterUnconnected = ({
                 onKeyUp={onKeyUp}
                 value={filterText}
                 data-test="search-dashboard-input"
+                ref={onFocusInput}
             />
             {filterText && <ClearButton onClear={clearDashboardsFilter} />}
         </div>